refactor: paperless instead of document API
