+++ js/base.js
@@ -1,6 +1,5 @@
 import { API as BaseAPI, mix, load, conf, verify } from "yonius";
 
-import { DocumentAPI } from "./document";
 import { LocatorAPI } from "./locator";
 import { PaperlessAPI } from "./paperless";
 import { PickupAPI } from "./pickup";
@@ -13,7 +12,6 @@ const API_VERSION = "v1";
 const GRANT_TYPE = "client_credentials";
 
 export class API extends mix(BaseAPI).with(
-    DocumentAPI,
     LocatorAPI,
     PaperlessAPI,
     PickupAPI,
