+++ js/index.js
@@ -1,5 +1,4 @@
 export * from "./base";
-export * from "./document";
 export * from "./locator";
 export * from "./paperless";
 export * from "./pickup";
